netology_pd_diplom/requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.0
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
drf-spectacular==0.26.2
pytest-django==4.5.2
pytest~=7.3.1
model_bakery==1.11.0